+++ demo/storage-client/index.ts
@@ -1,22 +1,20 @@
 import {AlwatrStorageClient} from '@alwatr/storage-client';
 
-import type {DocumentObject} from '@alwatr/storage-client';
+import type {AlwatrDocumentObject} from '@alwatr/storage-client';
 
-interface User extends DocumentObject {
+interface User extends AlwatrDocumentObject {
   fname: string;
   lname: string;
   email: string;
   token?: string;
 }
 
-const token = process.env.TOKEN;
-if (token == null) {
-  throw new Error('token_not_defined');
-}
+const token = process.env.TOKEN ?? 'YOUR_SECRET_TOKEN';
 
 const db = new AlwatrStorageClient<User>({
   name: 'user-list',
-  host: 'http://127.0.0.1:80',
+  host: '127.0.0.1',
+  port: 9000,
   token,
 });
 
@@ -27,7 +25,7 @@ try {
   console.log('ali found: %o', ali);
   /**
    * {
-   *   _id: 'alimd',
+   *   id: 'alimd',
    *   fname: 'Ali',
    *   lname: 'MM',
    *   email: 'i@ali.md',
@@ -39,9 +37,8 @@ try {
 catch (err) {
   if ((err as Error).message === 'document_not_found') {
     console.log('ali not found');
-    ali = {
-      _id: 'alimd',
-      _updatedBy: 'demo',
+    const ali = {
+      id: 'alimd',
       fname: 'Ali',
       lname: 'Mihandoost',
       email: 'ali@mihandoost.com',
@@ -54,8 +51,7 @@ catch (err) {
 }
 
 await db.set({
-  _id: 'fmd',
-  _updatedBy: 'demo',
+  id: 'fmd',
   fname: 'Fatemeh',
   lname: 'Mihandoost',
   email: 'Fatemeh@mihandoost.com',
@@ -65,8 +61,8 @@ await db.set({
 console.log('has \'alimd\': %o', await db.has('alimd'));
 console.log('keys: %o', await db.keys());
 console.log('getAll: %o', await db.getAll());
-console.log('delete: %o', await db.delete('alimd'));
-console.log('delete: %o', await db.delete('fmd'));
+await db.delete('alimd');
+await db.delete('fmd');
 
 try {
   await db.delete('abcd');