+++ packages/core/storage-client/src/type.ts
@@ -1,5 +1,3 @@
-import type {DocumentObject} from '@alwatr/storage-engine';
-
 export type AlwatrStorageClientConfig = {
   /**
    * Storage name (like database name).
@@ -7,9 +5,14 @@ export type AlwatrStorageClientConfig = {
   name: string;
 
   /**
-   * Storage server host name (URL).
+   * Storage server host name.
    */
   host: string;
+
+  /**
+   * Storage server port number.
+   */
+  port: number;
 
   /**
    * Storage server token (like database password).
@@ -32,20 +35,3 @@ export type AlwatrStorageClientConfig = {
    */
   debug?: boolean;
 };
-
-type ServerResponseFailed = {
-  ok: false;
-  statusCode: number;
-  errorCode: string;
-  data?: Record<string, unknown>;
-};
-
-type ServerResponseSuccess<DataType> = {
-  ok: true;
-  statusCode?: number;
-  data: DataType;
-};
-
-export type ServerResponse<DataType extends Record<string, unknown> = DocumentObject> =
-  | ServerResponseSuccess<DataType>
-  | ServerResponseFailed;